+++ tests/models.py
@@ -1,6 +1,8 @@
 """Test models."""
 
-from pydantic import BaseModel
+from typing import Annotated
+
+from pydantic import BaseModel, PlainSerializer
 
 
 class BaseDeezerGWResponse(BaseModel):
@@ -13,17 +15,17 @@ class BaseDeezerGWResponse(BaseModel):
 class DeezerSong(BaseModel):
     """Deezer API Song."""
 
-    SNG_ID: int
+    SNG_ID: Annotated[int, PlainSerializer(str)]
     TRACK_TOKEN: str
-    DURATION: int
+    DURATION: Annotated[int, PlainSerializer(str)]
     ART_NAME: str
     SNG_TITLE: str
     VERSION: str | None = None
     ALB_TITLE: str
     ALB_PICTURE: str
-    FILESIZE_MP3_128: int
-    FILESIZE_MP3_320: int
-    FILESIZE_FLAC: int
+    FILESIZE_MP3_128: Annotated[int, PlainSerializer(str)]
+    FILESIZE_MP3_320: Annotated[int, PlainSerializer(str)]
+    FILESIZE_FLAC: Annotated[int, PlainSerializer(str)]
 
 
 class DeezerSongResponse(BaseDeezerGWResponse):